homework 3 task 1 package renamed
