+++ homework3/pack_task1/module_task1.py
@@ -1,0 +1,57 @@
+# -*- coding: utf-8 -*-
+"""In previous homework task 4, you wrote a cache function that remembers other function
+output value. Modify it to be a parametrized decorator, so that the following code:
+
+@cache(times=3)
+def some_function():
+    pass
+Would give out cached value up to times number only. Example:
+
+@cache(times=2)
+def f():
+    return input('? ')   # careful with input() in python2, use raw_input() instead
+
+f()
+? 1
+'1'
+f()     # will remember previous value
+'1'
+f()     # but use it up to two times only
+'1'
+f()
+? 2
+'2'
+"""
+
+
+import time
+
+
+def clock(func):
+    def clocked(*args, **kwargs):
+        t0 = time.time()
+
+        result = func(*args, **kwargs)  # вызов декорированной функции
+
+        elapsed = time.time() - t0
+        name = func.__name__
+        arg_1st = []
+        if args:
+            arg_1st.append(', '.join(repr(arg) for arg in args))
+        if kwargs:
+            pairs = ['%s=%r' % (k, w) for k, w in sorted(kwargs.items())]
+            arg_1st.append(', '.join(pairs))
+        arg_str = ', '.join(arg_1st)
+        print('[%0.8fs] %s(%s) -> %r' % (elapsed, name, arg_str, result))
+        return result
+    return clocked
+
+
+@clock
+def fib(n):
+    if n < 2:
+        return n
+    return fib(n-2) + fib(n-1)
+
+
+print('fib(20) =', fib(20))
